Add the functionality to replace the cancel button with the first item
This includes using that item's icon, title and handler on the floating button while the list is open. When only one item is provided and the cancel button is not used, then handle that item's handler without opening the menu.

This commit is incomplete as it only handles the animations for the pop animation type.