Add the all animations and logic for a list to appear without a cancel button
When there is no cancel button (hasCancelButton = false) the list that appears will make the main Floaty button essentially the first item. All animations are supported and it will respond to tapping it if there is a handler on the item.

If hasCancelButton = false and there is only one item, then no list will appear but it will be as if the first item was selected.